Move DataSelector into querying
