+++ frontend/src/metabase/admin/datamodel/hoc/FilteredToUrlTable.tsx
@@ -7,7 +7,7 @@ import { t } from "ttag";
 import { FieldSet } from "metabase/common/components/FieldSet";
 import CS from "metabase/css/core/index.css";
 import { Tables } from "metabase/entities/tables";
-import { DatabaseSchemaAndTableDataSelector } from "metabase/query_builder/components/DataSelector";
+import { DatabaseSchemaAndTableDataSelector } from "metabase/querying/common/components/DataSelector";
 import type { State } from "metabase/redux/store";
 import { Icon } from "metabase/ui";
 import { connect } from "metabase/utils/redux";
