+++ frontend/src/metabase/query_builder/components/template_tags/TagEditorParamParts/TableMappingSelect.tsx
@@ -2,7 +2,7 @@ import { useId } from "react";
 import { t } from "ttag";
 
 import { LabelWithInfo } from "metabase/common/components/LabelWithInfo";
-import { SchemaAndTableDataSelector } from "metabase/query_builder/components/DataSelector";
+import { SchemaAndTableDataSelector } from "metabase/querying/common/components/DataSelector";
 import { Group, Switch } from "metabase/ui";
 import type Database from "metabase-lib/v1/metadata/Database";
 import type { TableId, TemplateTag } from "metabase-types/api";
