+++ frontend/src/metabase/querying/common/components/DataSelector/DataSelectorSectionHeader/DataSelectorSectionHeader.tsx
