+++ frontend/src/metabase/querying/common/components/DataSelector/DataSelectorSectionHeader/index.ts
