+++ frontend/src/metabase/querying/common/components/DataSelector/DataSelectorTablePicker/DataSelectorTablePicker.tsx
