+++ frontend/src/metabase/querying/common/components/DataSelector/DataSelectorTablePicker/index.ts
