+++ frontend/src/metabase/querying/common/components/DataSelector/saved-entity-picker/SavedEntityPicker.module.css
